+++ test/clj/matcher_combinators/matchers_test.clj
@@ -9,8 +9,8 @@
             [matcher-combinators.matchers :as m]
             [matcher-combinators.result :as result]
             [matcher-combinators.test :refer [match?]]
-            [matcher-combinators.test-helpers :as test-helpers :refer [abs-value-matcher]])
-  (:import [matcher_combinators.model Mismatch Missing InvalidMatcherType]))
+            [matcher-combinators.test-helpers :as test-helpers :refer [no-match? abs-value-matcher]])
+  (:import [matcher_combinators.model Mismatch Missing InvalidMatcherContext InvalidMatcherType]))
 
 (defn any? [_x] true)
 
@@ -42,6 +42,8 @@
   (instance? Missing actual))
 (defn invalid-type? [actual]
   (instance? InvalidMatcherType actual))
+(defn invalid-matcher-context? [actual]
+  (instance? InvalidMatcherContext actual))
 
 (defn one-mismatch? [mismatch-list]
   (= 1 (count (filter #(or (mismatch? %) (missing? %)) mismatch-list))))
@@ -243,9 +245,6 @@
     (is (= m/regex
            (m/matcher-for #"abc")))))
 
-(defn no-match? [expected actual]
-  (not (c/indicates-match? (c/match expected actual))))
-
 (deftest match-with-matcher
   (testing "processes overrides in order"
     (let [matcher (m/match-with [pos? abs-value-matcher
@@ -296,9 +295,33 @@
            #{1 2}))
 
       (is (match?
+            (m/match-with [set? m/embeds]
+                          #{(m/pred odd?)})
+            #{1 2}))
+      (is (match?
            (m/match-with [set? m/embeds]
                          #{odd?})
            #{1 2}))))
+
+  (testing "non-nested matchers"
+    (testing "irrelevant match-with doesn't affect results"
+      (is (match? (m/match-with [map? m/equals]
+                                [:key (m/regex #"valu*")])
+                  [:key "value"]))
+      (is (match? [:key (m/regex #"valu*")]
+                  [:key "value"]))
+
+      (is (match? (m/match-with [map? m/equals]
+                                [:key (m/pred even?)])
+                  [:key 2]))
+      (is (match? [:key (m/pred even?)]
+                  [:key 2]))
+
+      (is (match? (m/match-with [vector? m/equals]
+                                {:key (m/regex #"value")})
+                  {:key "value"}))
+      (is (match? {:key (m/regex #"value")}
+                  {:key "value"}))))
 
   (testing "multiple scopes"
     (let [expected
@@ -431,6 +454,11 @@
                ::result/value  {:expected "seq-of expects a non-empty sequence" :actual []}
                ::result/weight number?}
         (c/match (m/seq-of integer?) [])))
+  (testing "`seq-of` + `absent` fails because absent should only be used in maps"
+    (is (match? {::result/type  :mismatch
+                 ::result/value [invalid-matcher-context?]}
+                (c/match (m/seq-of m/absent)
+                         [1]))))
   (is (match? (m/seq-of {:name string? :id (partial instance? java.util.UUID)})
               [{:name "Michael"
                 :id    #uuid "c70e35eb-9eb6-4e3d-b5da-1f7f80932db9"}])))