+++ test/cljs/matcher_combinators/cljs_example_test.cljs
@@ -1,14 +1,13 @@
 (ns matcher-combinators.cljs-example-test
-  (:require [clojure.test :refer [deftest testing is are]]
+  (:require [clojure.test :refer [deftest testing is]]
             [clojure.test.check.generators :as gen]
             [clojure.test.check.properties :as prop]
             [clojure.test.check.clojure-test :refer [defspec]]
             [matcher-combinators.standalone :as standalone]
             [matcher-combinators.parser]
             [matcher-combinators.matchers :as m]
-            [matcher-combinators.core :as c]
-            [matcher-combinators.test]
-            [matcher-combinators.test-helpers :as helpers])
+            [matcher-combinators.test :refer [match? thrown-match?]]
+            [matcher-combinators.test-helpers :refer [abs-value-matcher no-match?]])
   (:import [goog.Uri]))
 
 (def gen-any-equatable
@@ -45,7 +44,12 @@
                            [2 3 1])))
   (testing "`(sort 2)` throws and causes a mismatch"
     (is (not (standalone/match? (m/via sort 2)
-                                2)))))
+                                2))))
+  (testing "via + match-with allows pre-processing `actual` before applying matching"
+    (is (match? (m/match-with
+                  [vector? (fn [expected] (m/via sort expected))]
+                  {:payloads [1 2 3]})
+                {:payloads (shuffle [3 2 1])}))))
 
 (deftest exception-matching
   (is (thrown-match? ExceptionInfo
@@ -54,6 +58,12 @@
 
 (deftest passing-match
   (is (match? {:a 2} {:a 2 :b 1})))
+
+(deftest pred-match
+  (is (match? #{odd?}
+              #{1}))
+  (is (match? #{(m/pred odd?)}
+              #{1})))
 
 (comment
   (deftest match?-no-actual-arg
@@ -68,3 +78,50 @@
     (testing "fails with nice message when you don't provide an `actual` arg to `thrown-match?`"
       (is (thrown-match? ExceptionInfo {:a 1})
           :in-wrong-place))))
+
+(deftest match-with-test
+  (testing "Example numeric test-case"
+    (is (match? (m/match-with [number? (m/within-delta 0.05)] 1) 0.99)))
+
+  (testing "maps"
+    (testing "passing case with equals override"
+      (is (match? (m/match-with [map? m/equals]
+                                {:a :b})
+                  {:a :b})))
+    (testing "failing case with equals override"
+      (is (no-match? (m/match-with [map? m/equals]
+                                   {:a :b})
+                     {:a :b :d :e})))
+    (testing "passing case multiple scopes"
+      (is (match?
+            {:o (m/match-with [map? m/equals]
+                              {:a
+                               (m/match-with [map? m/embeds]
+                                             {:b :c})})}
+            {:o {:a {:b :c :d :e}}
+             :p :q})))
+    (testing "using `absent` matcher"
+      (is (match? (m/match-with [map? m/equals]
+                                {:a m/absent
+                                 :b :c})
+                  {:b :c}))
+      (is (match? (m/match-with [map? m/embeds]
+                                {:a m/absent})
+                  {:b :c}))))
+
+  (testing "sets"
+    (is (match?
+          (m/match-with [set? m/embeds]
+                        #{1})
+          #{1 2}))
+
+    (is (match?
+          (m/match-with [set? m/embeds]
+                        #{(m/pred odd?)})
+          #{1 2})))
+
+  (let [matcher (m/match-with [pos? abs-value-matcher
+                               integer? m/equals]
+                              5)]
+    (is (match? matcher 5))
+    (is (match? matcher -5))))